// src/app/models/course.model.ts
export interface Course {
  id: number;             // 課程 ID
  title: string;          // 課程名稱
  description: string;    // 課程介紹
  company: string;        // 發布公司名稱
  price: number;          // 價格（例如 1200）
  imageUrl: string;       // 課程圖片網址
}